Move to using macros
Instead of having each function know how to handle it’s failure case-
allow the function to fail gracefully and use macros to handle errors
in a standard way.

This allows functions to be easily testable returning Result with Ok or
Err. Then we can use a macro to handle printing helpful user errors out
to the console in a standard way.