Don't split on every pane.
The previous logic made a command for a split in the window for every pane found. This means a window with 3 panes receives 3 splits, resulting in 4 panes.
If a window exists with a single pane and is split you have two panes. A split is needed for 1 less, than the amount of panes needed.