Remove named first window on session open
It seemed keen at first to name the first window open when a new session was started. The downside to this was it didn't respect the -c flag and became harder to set the default directory. It also meant the first defined window in the yaml file was treated different then every other window. This results in much code duplication and inconsistent window handling.

Instead opt for a method of giving the first window a random name, opening up all the user defined windows, then killing off the first. This means all windows are treated equally. The other issue with this method is the indexing of the remaining windows. It will offset the current users default window indexing. This can be addressed later on.