Better targeting when killing the first window.
Saw a bug where the default window created in the session is supposed to be killed it doesn't target the session properly.
This results in sending the kill window command to the last active session which could be any session. The window will not likely be found and cause an error.
As well as the default window would be left opened.