Fixes #7 No command provided for execution.
This was a bug where we always expected a value to be present to execute in Windows or Panes. There was conveniently an expected error message when it occured but this is still the wrong behaviour.
It is valid YAML to have a key with a None value. As well as to have Panes as an array where each item is None. Handle this gracefully.

Closes #7