Indexing problem
This refactoring was required to address the base-index issue. I was dealing with windows all the same by simply abandoning the first window as a tmp window. This had the unfortunite side effect of screwing with the base-index numbers.
To alleviate this I've opted to continue treating all windows the same while config matching, but to wait to set the session command until all window commands are created. Then we can remove the first window command in favour of a session command and put it back at the front of the stack for execution.

This eliminates the base-index problem as now all windows will be created at the user configurable base-index, indepedent of muxed.