issue-9 Windows the default focus to vi interupt the creation of the integration completion file.
To make that a session is finish being brought up, all integration tests touch a `/tmp/{project_name}.completed` file.
This is done through send-keys as a final command to the session. If the window default focuses to vi, it will write the touch of this file into a new vi buffer and will never execute.