issue-9 Select window and pane
Utilize the new commands to change the focus of the session to the topmost pane of the first window.
We're using the top most window as muxed currently has no concept of the pane indexes from a users config.
Top most is likely a good guestimate for the first pane of the window that is focues on.