issue-5 Shell out to the subcommand
Instead of using the native clap api for subcommands, manually check the first arg for named subcommands.
If `new` is found, shell out to muxednew.

Manually check for the existence of a call to --help. This is so the default clap interface won't print the main app help message and return. We can now check for a subcommand, and check for --help called on a subcommand, and return the appropriate help & usage.