feature(Custom Session name): Take a name setting
Allow the use of a name setting to be defined in the yaml config file. Implications
on this are checking to see if the session already exists. We now need to parse
the yaml file to find if it had a custom name, before we can check to see if the
file exists.