fix(Attach): Attach to named sessions with spaces
In the latest feature for named session (see #26) I created a bug. If the session
contained a space in the name it wouldn't attach to the session. This was caused
by only trying to attach to a truncated session name.

ex. `name: Brians Session` would fail trying to attach to a session `Brians`
that does not exist.