imp(Root paths): New window path reliability
When a new sessions is created with -c using the window path it sets the session
default path as well. This means new windows the user opens will be opened to
the wrong path. Instead set -c on the session as the root path, and navigate the
first window to the desird path directly. Similar to the old method.