fix: Improper passing of subcommands as project
Contributions by @coreyja to address a problem where subcommands that are called without a project name result in using the subcommand itself as a project name.
Instead check to ensure when using the default execution without a subcommand a project with the name of a subcommand isn't called accidentally. Do allow projects with the same name as subcommands but only while explicitly using the `load` subcommand.

Thanks @coreyja!

Closes #45, closes #46