imp: Support switching between sessions
This adds support to call muxed from within a running tmux session. You can call muxed simply to switch to another running session. Or call muxed to load a config, start a new session, and switch to it from within a running tmux session.
Previously it would have attempted to call tmux in a manner that would nest sessions instead of running them in parrallel. On the other hand if the desire is to nest sessions one simply needs to follow TMUX's suggestion and UNSET the $TMUX env variable.

Fixes: #49