Fix 'index out of bounds' in Tmux config parsing
I was getting this error
```
thread 'main' panicked at 'index out of bounds: the len is 1 but the index is 1', load/src/tmux/config.rs:25:35
```

From looking at the options tmux is printing on my machine I found the
issue was with the `pane-colours` option. Which didn't have a value in
my TMUX. This is while running a potentially pre-release version from
Homebrew

Just making it so we get an empty string if there is no value at least
gets around the panic.

Assuming `muxed` doesn't rely on `pane-colours` it _feels_ like it might
be a reasonable move. We could also not write `pane-colours` at all if
it doesn't have a value.