Add custom error types
The project had an unhealthy use of expect. It was easy as the bin runs
or doesn't. So throwing an expect was simple and not a detriment. But it
looks ugly. So add custom error types and handle errors better
throughout the application.

A little bit of additional function cleanup with naming, and following
new idioms.